# docs/swagger.yaml
definitions:
...
  comment.CommentCreateReq:
    properties:
      content:
        type: string
      contributionId:
        type: integer
    type: object
  comment.CommentRes:
    properties:
      content:
        type: string
      createdAt:
        type: string
      edited:
        type: boolean
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  comment.CommentUpdateReq:
    properties:
      content:
        type: string
    type: object
  comment.CursorPaginateComposition:
    properties:
      data:
        items:
          $ref: '#/definitions/comment.CommentRes'
        type: array
      next:
        type: string
    type: object
  contributesession.PaginateComposition:
    properties:
      currentPage:
        type: integer
      data:
        items:
          $ref: '#/definitions/contributesession.SessionRes'
        type: array
      lastPage:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
  contributesession.SessionCreateReq:
    properties:
      closureTime:
        type: string
      finalClosureTIme:
        type: string
      openTime:
        type: string
    type: object
  contributesession.SessionRes:
    properties:
      closureTime:
        type: string
      createdAt:
        type: string
      exportedAssets:
        type: string
      finalClosureTIme:
        type: string
      id:
        type: integer
      openTime:
        type: string
      updatedAt:
        type: string
    type: object
  contributesession.SessionUpdateReq:
    properties:
      closureTime:
        type: string
      finalClosureTIme:
        type: string
      openTime:
        type: string
    type: object
...
  user.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      facultyId:
        type: integer
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  title: "123"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: List comments
      parameters:
      - in: query
        name: contributionId
        required: true
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: next
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CursorPaginateComposition'
      security:
      - ApiKeyAuth: []
      summary: List comments
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a comment
      parameters:
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/comment.CommentCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentRes'
      security:
      - ApiKeyAuth: []
      summary: Create a comment
      tags:
      - Comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
    get:
      consumes:
      - application/json
      description: get comment by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentRes'
      security:
      - ApiKeyAuth: []
      summary: Show a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a comment
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/comment.CommentUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/comment.CommentRes'
      security:
      - ApiKeyAuth: []
      summary: Update a comment
      tags:
      - Comments
  /contribute-sessions:
    get:
      consumes:
      - application/json
      description: List Contribute Sessions
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contributesession.PaginateComposition'
      security:
      - ApiKeyAuth: []
      summary: List Contribute Sessions
      tags:
      - Contribute Sessions
    post:
      consumes:
      - application/json
      description: Create a Contribute Session
      parameters:
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contributesession.SessionCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contributesession.SessionRes'
      security:
      - ApiKeyAuth: []
      summary: Create a Contribute Session
      tags:
      - Contribute Sessions
  /contribute-sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Contribute Session
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a Contribute Session
      tags:
      - Contribute Sessions
    get:
      consumes:
      - application/json
      description: get Contribute Session by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contributesession.SessionRes'
      security:
      - ApiKeyAuth: []
      summary: Show a Contribute Session
      tags:
      - Contribute Sessions
    put:
      consumes:
      - application/json
      description: Update a Contribute Session
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contributesession.SessionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contributesession.SessionRes'
      security:
      - ApiKeyAuth: []
      summary: Update a Contribute Session
      tags:
      - Contribute Sessions
  /contribute-sessions/{id}/export:
    post:
      consumes:
      - application/json
      description: Export a Contribute Session
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Export a Contribute Session
      tags:
      - Contribute Sessions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"